Misc small cleanup

+++ include/d/d_drawlist.h
@@ -238,7 +238,7 @@ public:
     void setScissor(f32, f32, f32, f32);
 
     view_port_class* getViewPort() { return &mViewport; }
-    s8 getCameraID() { return mCameraID; }
+    int getCameraID() { return mCameraID; }
     void setCameraID(int id) { mCameraID = id; }
     void setMode(int mode) { mMode = mode; }
